==> research/04_fetch_es_futures_overnight.ipynb <==
import yfinance as yf 
import pandas as pd 
import os 
from datetime import datetime

data_base_path = "/Users/kennethzhang/Desktop/VolAuction/data"
auction_quotes = pd.read_parquet(data_base_path + "/auction_daily_quotes.parquet")
auction_dates = pd.to_datetime(auction_quotes["date"]).dt.tz_localize(None)

es_futures_df = pd.read_csv(data_base_path + "/es_futures_1m_all.csv", parse_dates=True, index_col=0)
es_futures_df.index = pd.to_datetime(es_futures_df.index)
es_futures_df.index = es_futures_df.index.tz_convert("America/New_York") 

es_futures_df

def get_es_slice(auction_date):
    t_minus_1 = auction_date - pd.Timedelta(days=1)
    start = pd.Timestamp(f"{t_minus_1.date()} 16:00", tz="America/New_York")
    end = pd.Timestamp(f"{auction_date.date()} 09:30", tz="America/New_York")
    return es_futures_df.loc[(es_futures_df.index >= start) & (es_futures_df.index <= end)].copy()

test_date = pd.Timestamp('2025-03-25', tz="America/New_York")
es_window = get_es_slice(test_date)

es_window
